refactor(oci): cut out to a separate file
